requirements.txt
setuptools
requests
pyaml
paramiko
git+https://github.com/sonata-nfv/tng-cli.git#egg=tngcli
git+https://github.com/sonata-nfv/tng-sdk-package.git#egg=tngpkg

